+++ cmd/cleat/template_proxy_process_windows_test.go
@@ -1,0 +1,14 @@
+//go:build windows
+
+package main
+
+import "os/exec"
+
+func inOwnProcessGroup(cmd *exec.Cmd) {}
+
+func killProcessGroup(cmd *exec.Cmd) {
+	if cmd.Process != nil {
+		_ = cmd.Process.Kill()
+		_, _ = cmd.Process.Wait()
+	}
+}